win32: Updated the VS2022 solution and modernized the dialog system look
Ported all of the changes we've made in Visual Studio 6 to our 2022
solution and updated the necessary files to get the "modern" look and
feel in application.

+++ win32/src/Application.cpp
@@ -10,6 +10,7 @@
 #include "AboutDialog.h"
 #include "SendFileDialog.h"
 
+// Common definitions.
 #define MAX_LOADSTRING 100
 
 // Global variables.
@@ -71,7 +72,7 @@ int APIENTRY wWinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 	}
 
 	// Clean up.
-	return TerminateInstance(hInstance, msg.wParam);
+	return TerminateInstance(hInstance, (int)msg.wParam);
 }
 
 /**
